In migration, if old and new both invalid, use new schema error
With a schema like this:

    class Old(Document):
        class __mongometa__:
            name = 'c'
        a = Field(str)

    class New(Document):
        class __mongometa__:
            name = 'c'
            version_of = Old
            def migrate(x):
                if not isinstance(x[a], list):
                x[a] = [x[a]]
                return x
        a = Field([str])
        b = Field(int)

Validating the document {'a': ['x'], 'b', 'y'} will either complain
about a not being a string or b being an extra key when the actual
error is that b is not an int.  This occurs because in Python 2, if
an exception is raised in an exception handler, the original exception
is lost.

This fixes the bug by catching the inner exception and re-raising the
outer exception in the validate function.