odmsession.update_if_not_modified: Properly pass upsert
Previsouly, when calling update_if_not_modified the call would fail and
raise an TypeError exception.  The problem was the `update` function
signature accepts only three positional arguments but we were attempting
to pass four.

The simple fix here is to pass `upsert`, the fourth positional
paramater, as a keyword argument.

In addition to this fix I have removed a useless if-statement.